Update backend/java/src/main/java/ai/snippetquiz/core_service/repository/QuizRepository.java
Method parameter type inconsistency: This method uses Integer id while the other similar method at line 20 uses Long id. Since the entity uses Long for the ID field, this should be Long id for consistency.

+++ backend/java/src/main/java/ai/snippetquiz/core_service/repository/QuizRepository.java
@@ -29,5 +29,5 @@ public interface QuizRepository extends JpaRepository<Quiz, Long> {
     @Query("SELECT q FROM Quiz q LEFT JOIN FETCH q.quizQuestions WHERE q.id = :id AND q.userId = :userId")
     Optional<Quiz> findByIdAndUserIdWithQuestions(@Param("id") Long id, @Param("userId") UUID userId);
 
-    Optional<Quiz> findByIdAndUserId(Integer id, UUID userId);
+    Optional<Quiz> findByIdAndUserId(Long id, UUID userId);
 }
